+++ lib/src/verify/header_body.rs
@@ -195,13 +195,16 @@ pub enum Error {
     NewRuntimeCompilationError(host::NewErr),
     /// Block being verified has erased the `:code` key from the storage.
     CodeKeyErased,
+    /// Parent storage has an empty `:code` key.
+    ///
+    /// > **Note**: This indicates that the parent block is invalid. This error is likely caused
+    /// >           by some kind of internal error or failed assumption somewhere in the API user's
+    /// >           code.
+    #[display(fmt = "Parent storage has an empty `:code` key")]
+    ParentCodeEmpty,
     /// Block has modified the `:heappages` key in a way that fails to parse.
     #[display(fmt = "Block has modified `:heappages` key in invalid way: {_0}")]
     HeapPagesParseError(executor::InvalidHeapPagesError),
-    /// Block has modified the `:heappages` key without modifying the `:code` key. This isn't
-    /// supported by smoldot.
-    // TODO: this is something that we should support but don't because it's annoying to implement and is clearly not worth the effort
-    HeapPagesOnlyModification,
 }
 
 /// Verifies whether a block is valid.
@@ -309,7 +312,7 @@ pub fn verify(
     // The first parameter of these two runtime functions is the same: a SCALE-encoded
     // `(header, body)` where `body` is a `Vec<Extrinsic>`. We perform the encoding ahead of time
     // in order to re-use it later for the second call.
-    let block_parameter = {
+    let execute_block_parameters = {
         // Consensus engines add a seal at the end of the digest logs. This seal is guaranteed to
         // be the last item. We need to remove it before we can verify the unsealed header.
         let mut unsealed_header = config.block_header.clone();
@@ -355,7 +358,7 @@ pub fn verify(
                         .chain(value_and_len.map(either::Right))
                 });
 
-                [either::Left(&block_parameter), either::Right(len)]
+                [either::Left(&execute_block_parameters), either::Right(len)]
                     .into_iter()
                     .map(either::Left)
                     .chain(encoded_list.map(either::Right))
@@ -375,7 +378,9 @@ pub fn verify(
 
     VerifyInner {
         inner: check_inherents_process,
-        execution_not_started: Some(block_parameter),
+        phase: VerifyInnerPhase::CheckInherents {
+            execute_block_parameters,
+        },
         consensus_success,
     }
     .run()
@@ -404,23 +409,36 @@ pub enum Verify {
 
 struct VerifyInner {
     inner: runtime_host::RuntimeHostVm,
-    /// If `Some`, then we are currently checking inherents, and this field contains the parameter
-    /// to later pass when invoking `Core_execute_block`. If `None`, then we are currently
-    /// executing the block.
-    execution_not_started: Option<Vec<u8>>,
+    phase: VerifyInnerPhase,
     consensus_success: SuccessConsensus,
+}
+
+enum VerifyInnerPhase {
+    CheckInherents {
+        /// Parameter to later pass when invoking `Core_execute_block`.
+        execute_block_parameters: Vec<u8>,
+    },
+    ExecuteBlock {
+        /// If the block modifies the `:heappages` but not the `:code`, we will need to fetch the
+        /// parent's `:code` in order to compile the new runtime. If that is the case, it will
+        /// be stored here.
+        parent_code: Option<Option<Vec<u8>>>,
+    },
 }
 
 impl VerifyInner {
     fn run(mut self) -> Verify {
         loop {
-            match self.inner {
-                runtime_host::RuntimeHostVm::Finished(Err(err)) => {
+            match (self.inner, self.phase) {
+                (runtime_host::RuntimeHostVm::Finished(Err(err)), _) => {
                     break Verify::Finished(Err((Error::WasmVm(err.detail), err.prototype)))
                 }
-                runtime_host::RuntimeHostVm::Finished(Ok(success))
-                    if self.execution_not_started.is_some() =>
-                {
+                (
+                    runtime_host::RuntimeHostVm::Finished(Ok(success)),
+                    VerifyInnerPhase::CheckInherents {
+                        execute_block_parameters,
+                    },
+                ) => {
                     // Check the output of the `BlockBuilder_check_inherents` runtime call.
                     let check_inherents_result =
                         check_check_inherents_output(success.virtual_machine.value().as_ref());
@@ -436,7 +454,7 @@ impl VerifyInner {
                         let vm = runtime_host::run(runtime_host::Config {
                             virtual_machine: success.virtual_machine.into_prototype(),
                             function_to_call: "Core_execute_block",
-                            parameter: iter::once(&self.execution_not_started.as_ref().unwrap()),
+                            parameter: iter::once(&execute_block_parameters),
                             storage_main_trie_changes: success.storage_main_trie_changes,
                             offchain_storage_changes: success.offchain_storage_changes,
                             max_log_level: 0,
@@ -452,11 +470,15 @@ impl VerifyInner {
 
                     self = VerifyInner {
                         consensus_success: self.consensus_success,
-                        execution_not_started: None,
+                        phase: VerifyInnerPhase::ExecuteBlock { parent_code: None },
                         inner: import_process,
                     };
                 }
-                runtime_host::RuntimeHostVm::Finished(Ok(success)) => {
+
+                (
+                    runtime_host::RuntimeHostVm::Finished(Ok(success)),
+                    VerifyInnerPhase::ExecuteBlock { parent_code },
+                ) => {
                     if !success.virtual_machine.value().as_ref().is_empty() {
                         return Verify::Finished(Err((
                             Error::NonEmptyOutput,
@@ -466,24 +488,32 @@ impl VerifyInner {
 
                     match (
                         success.storage_main_trie_changes.diff_get(&b":code"[..]),
+                        parent_code,
                         success
                             .storage_main_trie_changes
                             .diff_get(&b":heappages"[..]),
                     ) {
-                        (None, None) => {}
-                        (Some((None, ())), _) => {
+                        (None, _, None) => {}
+                        (Some((None, ())), _, _) => {
                             return Verify::Finished(Err((
                                 Error::CodeKeyErased,
                                 success.virtual_machine.into_prototype(),
                             )))
                         }
-                        (None, Some(_)) => {
+                        (None, None, Some(_)) => {
+                            break Verify::StorageGet(StorageGet {
+                                inner: StorageGetInner::ParentCode { success },
+                                consensus_success: self.consensus_success,
+                            })
+                        }
+                        (None, Some(None), _) => {
                             return Verify::Finished(Err((
-                                Error::HeapPagesOnlyModification,
+                                Error::ParentCodeEmpty,
                                 success.virtual_machine.into_prototype(),
                             )))
                         }
-                        (Some((Some(_code), ())), heap_pages) => {
+                        (Some((Some(_), ())), parent_code, heap_pages)
+                        | (_, parent_code @ Some(Some(_)), heap_pages) => {
                             let parent_runtime = success.virtual_machine.into_prototype();
 
                             let heap_pages = match heap_pages {
@@ -505,6 +535,7 @@ impl VerifyInner {
                                 consensus_success: self.consensus_success,
                                 parent_runtime,
                                 heap_pages,
+                                parent_code,
                                 logs: success.logs,
                                 offchain_storage_changes: success.offchain_storage_changes,
                                 storage_main_trie_changes: success.storage_main_trie_changes,
@@ -523,29 +554,30 @@ impl VerifyInner {
                         logs: success.logs,
                     }));
                 }
-                runtime_host::RuntimeHostVm::StorageGet(inner) => {
+
+                (runtime_host::RuntimeHostVm::StorageGet(inner), phase) => {
                     break Verify::StorageGet(StorageGet {
-                        inner,
-                        execution_not_started: self.execution_not_started,
+                        inner: StorageGetInner::Execution { inner, phase },
                         consensus_success: self.consensus_success,
                     })
                 }
-                runtime_host::RuntimeHostVm::PrefixKeys(inner) => {
+                (runtime_host::RuntimeHostVm::PrefixKeys(inner), phase) => {
                     break Verify::StoragePrefixKeys(StoragePrefixKeys {
                         inner,
-                        execution_not_started: self.execution_not_started,
+                        phase,
                         consensus_success: self.consensus_success,
                     })
                 }
-                runtime_host::RuntimeHostVm::NextKey(inner) => {
+                (runtime_host::RuntimeHostVm::NextKey(inner), phase) => {
                     break Verify::StorageNextKey(StorageNextKey {
                         inner,
-                        execution_not_started: self.execution_not_started,
+                        phase,
                         consensus_success: self.consensus_success,
                     })
                 }
-                runtime_host::RuntimeHostVm::SignatureVerification(sig) => {
+                (runtime_host::RuntimeHostVm::SignatureVerification(sig), phase) => {
                     self.inner = sig.verify_and_resume();
+                    self.phase = phase;
                 }
             }
         }
@@ -555,16 +587,28 @@ impl VerifyInner {
 /// Loading a storage value is required in order to continue.
 #[must_use]
 pub struct StorageGet {
-    inner: runtime_host::StorageGet,
-    /// See [`VerifyInner::execution_not_started`].
-    execution_not_started: Option<Vec<u8>>,
+    inner: StorageGetInner,
     consensus_success: SuccessConsensus,
+}
+
+enum StorageGetInner {
+    Execution {
+        inner: runtime_host::StorageGet,
+        /// See [`VerifyInner::phase`].
+        phase: VerifyInnerPhase,
+    },
+    ParentCode {
+        success: runtime_host::Success,
+    },
 }
 
 impl StorageGet {
     /// Returns the key whose value must be passed to [`StorageGet::inject_value`].
     pub fn key(&'_ self) -> impl AsRef<[u8]> + '_ {
-        self.inner.key()
+        match &self.inner {
+            StorageGetInner::Execution { inner, .. } => either::Left(inner.key()),
+            StorageGetInner::ParentCode { .. } => either::Right(b":code"),
+        }
     }
 
     /// Injects the corresponding storage value.
@@ -572,12 +616,27 @@ impl StorageGet {
         self,
         value: Option<(impl Iterator<Item = impl AsRef<[u8]>>, TrieEntryVersion)>,
     ) -> Verify {
-        VerifyInner {
-            inner: self.inner.inject_value(value),
-            execution_not_started: self.execution_not_started,
-            consensus_success: self.consensus_success,
+        match self.inner {
+            StorageGetInner::Execution { inner, phase } => VerifyInner {
+                inner: inner.inject_value(value),
+                phase,
+                consensus_success: self.consensus_success,
+            }
+            .run(),
+            StorageGetInner::ParentCode { success } => VerifyInner {
+                inner: runtime_host::RuntimeHostVm::Finished(Ok(success)),
+                phase: VerifyInnerPhase::ExecuteBlock {
+                    parent_code: Some(value.map(|(val_iter, _)| {
+                        val_iter.fold(Vec::new(), |mut a, b| {
+                            a.extend_from_slice(b.as_ref());
+                            a
+                        })
+                    })),
+                },
+                consensus_success: self.consensus_success,
+            }
+            .run(),
         }
-        .run()
     }
 }
 
@@ -585,8 +644,8 @@ impl StorageGet {
 #[must_use]
 pub struct StoragePrefixKeys {
     inner: runtime_host::PrefixKeys,
-    /// See [`VerifyInner::execution_not_started`].
-    execution_not_started: Option<Vec<u8>>,
+    /// See [`VerifyInner::phase`].
+    phase: VerifyInnerPhase,
     consensus_success: SuccessConsensus,
 }
 
@@ -600,7 +659,7 @@ impl StoragePrefixKeys {
     pub fn inject_keys_ordered(self, keys: impl Iterator<Item = impl AsRef<[u8]>>) -> Verify {
         VerifyInner {
             inner: self.inner.inject_keys_ordered(keys),
-            execution_not_started: self.execution_not_started,
+            phase: self.phase,
             consensus_success: self.consensus_success,
         }
         .run()
@@ -611,8 +670,8 @@ impl StoragePrefixKeys {
 #[must_use]
 pub struct StorageNextKey {
     inner: runtime_host::NextKey,
-    /// See [`VerifyInner::execution_not_started`].
-    execution_not_started: Option<Vec<u8>>,
+    /// See [`VerifyInner::phase`].
+    phase: VerifyInnerPhase,
     consensus_success: SuccessConsensus,
 }
 
@@ -649,7 +708,7 @@ impl StorageNextKey {
     pub fn inject_key(self, key: Option<impl Iterator<Item = Nibble>>) -> Verify {
         VerifyInner {
             inner: self.inner.inject_key(key),
-            execution_not_started: self.execution_not_started,
+            phase: self.phase,
             consensus_success: self.consensus_success,
         }
         .run()
@@ -668,19 +727,20 @@ pub struct RuntimeCompilation {
     offchain_storage_changes: storage_diff::TrieDiff,
     logs: String,
     heap_pages: vm::HeapPages,
+    parent_code: Option<Option<Vec<u8>>>,
     consensus_success: SuccessConsensus,
 }
 
 impl RuntimeCompilation {
     /// Performs the runtime compilation.
     pub fn build(self) -> Verify {
-        // A `RuntimeCompilation` object is built only if `:code` has been modified and to a
-        // specific value.
+        // A `RuntimeCompilation` object is built only if `:code` is available.
         let code = self
             .storage_main_trie_changes
             .diff_get(&b":code"[..])
+            .map(|(value, _)| value)
+            .or(self.parent_code.as_ref().map(|v| v.as_deref()))
             .unwrap()
-            .0
             .unwrap();
 
         let new_runtime = match host::HostVmPrototype::new(host::Config {